Moved ReviewsVM to it's own file #580
